Using `propagate` boolean to add callback delay of 30s
The S3Outposts bucket creation takes time. This results in the the next call failing...
For e.g. DELETE after a CREATE fails with the status "Bucket is in an invalid state".

In order to avoid this, I'm adding a 30s delay using the `propagate` boolean. This will give some time
for the bucket resource to stabilize.