Modifying version of package six to 1.15.0
An issue was identified with the version 1.14.0 of the package six that comes preinstalled with the
codebuild python environment. Version 1.14.0 is incompatible with some dependencies that were pulled in the
latest release of cfn-python-lint. This was causing the GitHub status checks to fail while creating a PR.

Upgrading the version of the package six to 1.15.0 to resolve this issue.