Adding retry logic to Bucket DeleteHandler (#30)
The service returns a 409 InvalidBucketState error when we try to Delete a bucket which is not
in the 'Active' state. This results in DeleteBucket failing in CFN if issued immediately after
a CreateBucket call. This also happens when the Create operation fails for other reasons when
done through CFN and CFN automatically tries to delete the bucket and rollback the stack. At such
times the DeleteBucket call fails since the Bucket was not in the correct state.

The fix is to check the error code and if its a 409 InvalidBucketState, we will keep issuing the
DeleteBucket call until it succeeds.

Further enhancements can be made - such as adding a limited number of retries - in case we see
situations where the Delete Retry operation from CFN never ends.

Successfully build using `mvn clean package`.