Adding retry logic to AccessPoint DeleteHandler (#31)
The service returns a 400 Invalid Request with an error message "Access Point is not in a state where it can be deleted"
when we try to delete an AccessPoint which is not in the expected 'Active' state. This results in DeleteAccessPoint failing.
When CFN issues a Delete as part of a rollback, if the Delete is issued immediately after a Create, it fails since the AccessPoint
is not in the 'Active' state.

The fix is to check the error code and error message and if they both match with what the service returns, we retry the Delete operation.

In the future, we can make additional enhancements such as retrying the operation only a limited number of times.

Successful build using `mvn clean package`.