Modifying ClientBuilder to use production endpoint + changing regex for Tag.Value
ClientBuilder is now configured to use the production endpoint for the S3Control and S3Outposts APIs.

Previously we were rejecting Tag.Value which had a prefix of `aws:`. This is incorrect behavior. While a Tag.Key cannot begin
with `aws:`, else it will be considered a systemTag, a Tag.Value can begin with `aws:`. This checkin fixes this behavior.
We now allow Tag.Value to begin with `aws:`.

Successful build using `mvn clean package`.